backend/requirements.txt
Flask==2.3.3
requests==2.31.0
pytz==2023.3
Flask-Cors==3.0.10
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
psycopg2-binary
python-dotenv==1.1.1
python-json-logger
prometheus_client